Create `EdgeEvaluator` and `NodeEvaluator` APIs
These APIs are for applying weights to nodes and edges in the graph, so
as to implement weighted PageRank as described in #476. As yet, we
are reasonably confident what the type signatures should be, but do not
know what functions will get the theoretical properties we want.

For now, I've implemented placeholder implementations. These are clearly
deficient in some ways, but at least allow us to start experimenting.

Paired with @wchargin.

Test plan:
No tests were added, as the code modifies prototype parts of our
codebase that are, themselves, not yet tested, and because we already
know these functions have bad semantics.

To ensure that this commit is reasonably correct, I experimented with
changing node/edge weights in the UI and verified that PageRank results
changed as expected.

(for emphasis): #476 has all the context on this PR.